update architecture api route

+++ database/migrations/v1/gourvernance/shareholder/2024_04_10_022310_create_shareholders_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->unsignedBigInteger('actions_number');
             $table->unsignedBigInteger('actions_encumbered');
             $table->unsignedBigInteger('actions_no_encumbered');
-            $table->integer('percentage')->nullable();
+            $table->decimal('percentage', 3, 2)->nullable();
 
             $table->uuid('created_by')->nullable();
             $table->foreign('created_by')->references('id')->on('users')->onDelete('cascade');
